pkg-package.yml: sign package
The user nobody is used to run makepkg as root is not allowed[1]. The
nobody is setup without an home directory set[2], and thus, the HOME
defaults to /[3].

GnuPG creates its home directory to ~/.gnupg if neither the environment
variable GNUPGHOME nor the option --homedir is given. The user nobody
does not have the permission to write to /.gnupg, causing gpg to fail.

makepkg uses gpg to verify the sources[4], to list the secret keys[5],
and to sign the packages using the gpg-agent[6].

This adds makepkg signature by adding a gpg wrapper script to the
in-PATH directory /usr/local/bin. The script sets an ephemeral GnuPG
home directory to /tmp/.gnupg to overcome the lack for permission to /
if run from user nobody, and it disables the use of the user-agent by
replacing the option --user-agent by the options needed to set the
passphrase from stdin.

Note: This hardcodes zstd compression.

According to gpg(1):

	Ephemeral home directories

	Sometimes you want to contain effects of some operation, for
	example you want to import a key to inspect it, but you do not
	want this key to be added to your keyring. In earlier versions
	of GnuPG, it was possible to specify alternate keyring files for
	both public and secret keys. In modern GnuPG versions, however,
	we changed how secret keys are stored in order to better protect
	secret key material, and it was not possible to preserve this interface.

	The preferred way to do this is to use ephemeral home
	directories. This technique works across all versions of GnuPG.

	Create a temporary directory, create (or copy) a configuration
	that meets your needs, make gpg use this directory either using
	the environment variable GNUPGHOME, or the option --homedir.
	GPGME supports this too on a per-context basis, by modifying the
	engine info of contexts. Now execute whatever operation you
	like, import and export key material as necessary. Once
	finished, you can delete the directory.  All GnuPG backend
	services that were started will detect this and shut down.

[1]: https://gitlab.archlinux.org/pacman/pacman/-/blob/v7.0.0/scripts/makepkg.sh.in?ref_type=tags#L1111-L1115
[2]: https://github.com/systemd/systemd/blob/v257.1/sysusers.d/basic.conf.in#L12-L14
[3]: https://github.com/systemd/systemd/blob/v257.1/man/sysusers.d.xml#L230-L231
[4]: https://gitlab.archlinux.org/pacman/pacman/-/blob/v7.0.0/scripts/libmakepkg/integrity/verify_signature.sh.in?ref_type=tags#L168
[5]: https://gitlab.archlinux.org/pacman/pacman/-/blob/v7.0.0/scripts/makepkg.sh.in?ref_type=tags#L1264
[6]: https://gitlab.archlinux.org/pacman/pacman/-/blob/v7.0.0/scripts/libmakepkg/integrity/generate_signature.sh.in?ref_type=tags#L37